New module for serial interrupt code.
